// src/api/openapi/models/ToolRequest.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type ToolRequest = {
    id?: string;
    tool_requested?: string;
    quantity_requested?: string;
    status?: string;
    requester_note?: string;
    client_note?: string;
    forwarded_to?: string;
    accepted_by?: Array<string>;
    contact?: string;
    farm_id?: string;
    createdat?: string;
    updatedat?: string;
    farm_name?: string;
    location?: string;
    description?: string;
    farm_head?: string;
    district?: string;
    size?: string;
    avatar?: string;
    cover_photo?: string;
    application_id?: string;
    is_archived?: boolean;
};